Reorganizo módulos de test
    refs:#94
