+++ series_tiempo_ar_api/apps/api/tests/pipeline_tests/__init__.py
@@ -1,0 +1,1 @@
+#! coding: utf-8
